refactor(superbpe): TokenizerTrainer class + replayable dataset factory
Replace the standalone train_tokenizer() function with a TokenizerTrainer
class that mirrors the model-side Trainer pattern:

- TokenizerTrainer(config, wandb_enabled=True) holds the validated config,
  builds an internal TokenizerWandbLogger and TokenizerMetricsTracker,
  and dispatches to _train_bpe / _train_superbpe.
- TokenizerWandbLogger is a thin wandb wrapper (init / log / finish only).
  It reads identity (project / run_name / group) from config.logging and
  records the full TrainConfig via config.to_dict().
- TokenizerMetricsTracker exposes exactly build_train_log_dict() and
  build_eval_log_dict() — both return {vocab_size, bytes_per_token}
  today, kept separate so eval-only metrics can land later.

Config schema: collapse the three tokenizer fields into DataConfig.
tokenizer_train_method + tokenizer_train_kwargs (a free-form dict
forwarded to the method), plus tokenizer_train_num_samples and
tokenizer_train_eval_every for script-level and curve-cadence
controls. wandb_enabled moves off LoggingConfig back to a CLI flag
(--no-wandb on scripts/train_tokenizer.py), matching scripts/train.py.

train() takes a Callable[[], Iterable[str]] factory instead of a
one-shot iterator: SuperBPE needs two corpus passes (stage 1 training
+ stage 2 re-encode for the merge loop), and we refuse to materialize
the whole stream in memory. The caller hands us something replayable;
internally we pull eval_texts from one fresh stream and build a
make_train_iter closure that produces additional fresh streams
skipping the eval slice (with a tiny-dataset fallback to eval_texts).

Also: BPE now logs the same (vocab_size, bytes_per_token) curve as
SuperBPE via post-hoc partial-tokenizer reconstruction; experiment
YAMLs migrate to the new schema; test_tokenizer.py shrinks to
load_tokenizer coverage and trainer tests move to test_tokenizer_trainer.py.
