+++ experiments/superbpe/run_train_tokenizers.sh
@@ -6,7 +6,5 @@ for cfg in experiments/superbpe/*.yaml; do
     name=$(basename "$cfg" .yaml)
     echo "==> Training $name"
     uv run python scripts/train_tokenizer.py --config "$cfg" \
-        --num_samples 100000 \
-        --wandb_project superbpe-reproduction \
         2>&1 | tee "logs/${name}_train.log"
 done
