+++ tests/fast/utils/test_config.py
@@ -196,3 +196,34 @@ model:
     p.write_text(yaml_content)
     with pytest.raises(ValueError, match="unknown attn_implementation"):
         load_config(str(p))
+
+
+# ==================== tokenizer-training fields on DataConfig ====================
+
+
+def test_dataconfig_tokenizer_train_defaults():
+    dc = DataConfig()
+    assert dc.tokenizer_train_method == "bpe"
+    assert dc.tokenizer_train_kwargs == {}
+
+
+def test_dataconfig_loads_superbpe_yaml(tmp_path):
+    yaml_content = """
+model:
+  vocab_size: 200000
+data:
+  dataset: openwebtext
+  tokenizer_path: tokenizers/superbpe_200k_t80k
+  tokenizer_train_method: superbpe
+  tokenizer_train_kwargs:
+    transition_size: 80000
+    max_superword_words: 4
+"""
+    p = tmp_path / "cfg.yaml"
+    p.write_text(yaml_content)
+    cfg = load_config(str(p))
+    assert cfg.data.tokenizer_train_method == "superbpe"
+    assert cfg.data.tokenizer_train_kwargs == {
+        "transition_size": 80000,
+        "max_superword_words": 4,
+    }
